+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofill.kt
@@ -1,6 +1,11 @@
 package com.mapbox.search.autofill
 
+import android.Manifest
+import com.mapbox.android.core.location.LocationEngine
+import com.mapbox.android.core.location.LocationEngineProvider
 import com.mapbox.geojson.Point
+import com.mapbox.search.base.BaseSearchSdkInitializer
+import com.mapbox.search.base.location.defaultLocationEngine
 
 /**
  * Main entrypoint to the Mapbox Autofill API. Performs forward or reverse geocoding requests to fetch addresses.
@@ -41,10 +46,26 @@ public interface AddressAutofill {
          *
          * @param accessToken [Mapbox Access Token](https://docs.mapbox.com/help/glossary/access-token/).
          *
+         * @param locationEngine The mechanism responsible for providing location approximations to the SDK.
+         * By default [LocationEngine] is retrieved from [LocationEngineProvider.getBestLocationEngine].
+         * Note that this class requires [Manifest.permission.ACCESS_COARSE_LOCATION] or
+         * [Manifest.permission.ACCESS_FINE_LOCATION] to work properly.
+         *
          * @return a new instance instance of [AddressAutofill].
          */
-        public fun create(accessToken: String): AddressAutofill {
-            return AddressAutofillImpl(AutofillSearchEngine.create(accessToken))
+        @JvmStatic
+        @JvmOverloads
+        public fun create(
+            accessToken: String,
+            locationEngine: LocationEngine = defaultLocationEngine(),
+        ): AddressAutofill {
+            return AddressAutofillImpl(
+                AutofillSearchEngine.create(
+                    accessToken = accessToken,
+                    app = BaseSearchSdkInitializer.app,
+                    locationEngine = locationEngine
+                )
+            )
         }
     }
 }