+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/location/LocationUtils.kt
@@ -1,10 +1,16 @@
 package com.mapbox.search.base.location
 
+import android.content.Context
+import com.mapbox.android.core.location.LocationEngineProvider
 import com.mapbox.geojson.BoundingBox
+import com.mapbox.search.base.BaseSearchSdkInitializer
 import kotlin.math.abs
 import kotlin.math.log2
 import kotlin.math.max
 import kotlin.math.min
+
+fun defaultLocationEngine(context: Context = BaseSearchSdkInitializer.app) =
+    LocationEngineProvider.getBestLocationEngine(context)
 
 fun calculateMapZoom(bbox: BoundingBox): Float {
     val eps = 1.0E-5